requirements.txt
dipy==1.11.0
matplotlib==3.10.3
nibabel==5.3.2
numpy==2.2.6
pytorch_warmup==0.2.0
PyYAML==6.0.2
scipy==1.15.3
torch==2.7.1
torchio==0.20.14
tqdm==4.67.1
swanlab
